Update on going

+++ S33M3Engines/S33M3CoreComponents/Cameras/FirstPersonCameraWithFocus.cs
@@ -87,7 +87,7 @@ namespace S33M3CoreComponents.Cameras
             }
         }
 
-        public override void VTSUpdate(double interpolationHd, float interpolationLd, long elapsedTime)
+        public override void VTSUpdate(double interpolationHd, float interpolationLd, float elapsedTime)
         {
             //Do interpolation on the value received at update time
             Vector3D.Lerp(ref _worldPosition.ValuePrev, ref _worldPosition.Value, interpolationHd, out _worldPosition.ValueInterp);
