+++ Utopia/Utopia/Entities/Renderer/GhostedEntityRenderer.cs
@@ -105,7 +105,7 @@ namespace Utopia.Entities.Renderer
             _voxelModelEffect = ToDispose(new HLSLVoxelModel(context.Device, ClientSettings.EffectPack + @"Entities\VoxelModel.hlsl", VertexVoxel.VertexDeclaration));
         }
 
-        public override void VTSUpdate(double interpolationHd, float interpolationLd, long elapsedTime)
+        public override void VTSUpdate(double interpolationHd, float interpolationLd, float elapsedTime)
         {
             if (Tool == null)
                 return;
